# [mypyc] Use native calls to singledispatch functions