Fixes mypyc/mypyc#881

This makes generated code use native calls to singledispatch functions whenever possible, which makes singledispatch functions faster.

## Benchmark results

This PR:
```
running sum_tree_singledispatch
..........
interpreted: 0.593835s (avg of 5 iterations; stdev 2.4%)
compiled:    0.044264s (avg of 5 iterations; stdev 2.1%)

compiled is 13.416x faster
```

Current master (4c90aa42e6a74f127956124aec94f55dbc46a281):
```
running sum_tree_singledispatch
..........
interpreted: 0.588636s (avg of 5 iterations; stdev 1.4%)
compiled:    0.059705s (avg of 5 iterations; stdev 1.5%)

compiled is 9.859x faster
```

## Test Plan

I added an irbuild test to make sure that we were using native calls to singledispatch functions.